Apply the same tests to all storage backends (#453)
