+++ taskchampion/src/storage/mod.rs
@@ -14,6 +14,9 @@ use crate::operation::Operation;
 use std::collections::HashMap;
 use uuid::Uuid;
 
+#[cfg(test)]
+mod test;
+
 mod config;
 mod inmemory;
 pub(crate) mod sqlite;
