+++ taskchampion/src/storage/sqlite.rs
@@ -464,6 +464,13 @@ mod test {
     use std::time::Duration;
     use tempfile::TempDir;
 
+    fn storage() -> Result<SqliteStorage> {
+        let tmp_dir = TempDir::new()?;
+        SqliteStorage::new(tmp_dir.path(), true)
+    }
+
+    crate::storage::test::storage_tests!(storage()?);
+
     /// Manually create a 0_7_0 db, as based on a dump from an actual (test) user.
     /// This is used to test in-place upgrading.
     fn create_0_7_0_db(path: &Path) -> Result<()> {
@@ -655,521 +662,6 @@ mod test {
                 num_ops += 1;
             }
             assert_eq!(num_ops, 15);
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn drop_transaction() -> Result<()> {
-        let tmp_dir = TempDir::new()?;
-        let mut storage = SqliteStorage::new(tmp_dir.path(), true)?;
-        let uuid1 = Uuid::new_v4();
-        let uuid2 = Uuid::new_v4();
-
-        {
-            let mut txn = storage.txn()?;
-            assert!(txn.create_task(uuid1)?);
-            txn.commit()?;
-        }
-
-        {
-            let mut txn = storage.txn()?;
-            assert!(txn.create_task(uuid2)?);
-            std::mem::drop(txn); // Unnecessary explicit drop of transaction
-        }
-
-        {
-            let mut txn = storage.txn()?;
-            let uuids = txn.all_task_uuids()?;
-
-            assert_eq!(uuids, [uuid1]);
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_create() -> Result<()> {
-        let tmp_dir = TempDir::new()?;
-        let mut storage = SqliteStorage::new(tmp_dir.path(), true)?;
-        let uuid = Uuid::new_v4();
-        {
-            let mut txn = storage.txn()?;
-            assert!(txn.create_task(uuid)?);
-            txn.commit()?;
-        }
-        {
-            let mut txn = storage.txn()?;
-            let task = txn.get_task(uuid)?;
-            assert_eq!(task, Some(taskmap_with(vec![])));
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn test_create_exists() -> Result<()> {
-        let tmp_dir = TempDir::new()?;
-        let mut storage = SqliteStorage::new(tmp_dir.path(), true)?;
-        let uuid = Uuid::new_v4();
-        {
-            let mut txn = storage.txn()?;
-            assert!(txn.create_task(uuid)?);
-            txn.commit()?;
-        }
-        {
-            let mut txn = storage.txn()?;
-            assert!(!txn.create_task(uuid)?);
-            txn.commit()?;
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn test_get_missing() -> Result<()> {
-        let tmp_dir = TempDir::new()?;
-        let mut storage = SqliteStorage::new(tmp_dir.path(), true)?;
-        let uuid = Uuid::new_v4();
-        {
-            let mut txn = storage.txn()?;
-            let task = txn.get_task(uuid)?;
-            assert_eq!(task, None);
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn test_set_task() -> Result<()> {
-        let tmp_dir = TempDir::new()?;
-        let mut storage = SqliteStorage::new(tmp_dir.path(), true)?;
-        let uuid = Uuid::new_v4();
-        {
-            let mut txn = storage.txn()?;
-            txn.set_task(uuid, taskmap_with(vec![("k".to_string(), "v".to_string())]))?;
-            txn.commit()?;
-        }
-        {
-            let mut txn = storage.txn()?;
-            let task = txn.get_task(uuid)?;
-            assert_eq!(
-                task,
-                Some(taskmap_with(vec![("k".to_string(), "v".to_string())]))
-            );
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn test_delete_task_missing() -> Result<()> {
-        let tmp_dir = TempDir::new()?;
-        let mut storage = SqliteStorage::new(tmp_dir.path(), true)?;
-        let uuid = Uuid::new_v4();
-        {
-            let mut txn = storage.txn()?;
-            assert!(!txn.delete_task(uuid)?);
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn test_delete_task_exists() -> Result<()> {
-        let tmp_dir = TempDir::new()?;
-        let mut storage = SqliteStorage::new(tmp_dir.path(), true)?;
-        let uuid = Uuid::new_v4();
-        {
-            let mut txn = storage.txn()?;
-            assert!(txn.create_task(uuid)?);
-            txn.commit()?;
-        }
-        {
-            let mut txn = storage.txn()?;
-            assert!(txn.delete_task(uuid)?);
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn test_all_tasks_empty() -> Result<()> {
-        let tmp_dir = TempDir::new()?;
-        let mut storage = SqliteStorage::new(tmp_dir.path(), true)?;
-        {
-            let mut txn = storage.txn()?;
-            let tasks = txn.all_tasks()?;
-            assert_eq!(tasks, vec![]);
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn test_all_tasks_and_uuids() -> Result<()> {
-        let tmp_dir = TempDir::new()?;
-        let mut storage = SqliteStorage::new(tmp_dir.path(), true)?;
-        let uuid1 = Uuid::new_v4();
-        let uuid2 = Uuid::new_v4();
-        {
-            let mut txn = storage.txn()?;
-            assert!(txn.create_task(uuid1)?);
-            txn.set_task(
-                uuid1,
-                taskmap_with(vec![("num".to_string(), "1".to_string())]),
-            )?;
-            assert!(txn.create_task(uuid2)?);
-            txn.set_task(
-                uuid2,
-                taskmap_with(vec![("num".to_string(), "2".to_string())]),
-            )?;
-            txn.commit()?;
-        }
-        {
-            let mut txn = storage.txn()?;
-            let mut tasks = txn.all_tasks()?;
-
-            // order is nondeterministic, so sort by uuid
-            tasks.sort_by(|a, b| a.0.cmp(&b.0));
-
-            let mut exp = vec![
-                (
-                    uuid1,
-                    taskmap_with(vec![("num".to_string(), "1".to_string())]),
-                ),
-                (
-                    uuid2,
-                    taskmap_with(vec![("num".to_string(), "2".to_string())]),
-                ),
-            ];
-            exp.sort_by(|a, b| a.0.cmp(&b.0));
-
-            assert_eq!(tasks, exp);
-        }
-        {
-            let mut txn = storage.txn()?;
-            let mut uuids = txn.all_task_uuids()?;
-            uuids.sort();
-
-            let mut exp = vec![uuid1, uuid2];
-            exp.sort();
-
-            assert_eq!(uuids, exp);
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn test_base_version_default() -> Result<()> {
-        let tmp_dir = TempDir::new()?;
-        let mut storage = SqliteStorage::new(tmp_dir.path(), true)?;
-        {
-            let mut txn = storage.txn()?;
-            assert_eq!(txn.base_version()?, DEFAULT_BASE_VERSION);
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn test_base_version_setting() -> Result<()> {
-        let tmp_dir = TempDir::new()?;
-        let mut storage = SqliteStorage::new(tmp_dir.path(), true)?;
-        let u = Uuid::new_v4();
-        {
-            let mut txn = storage.txn()?;
-            txn.set_base_version(u)?;
-            txn.commit()?;
-        }
-        {
-            let mut txn = storage.txn()?;
-            assert_eq!(txn.base_version()?, u);
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn test_operations() -> Result<()> {
-        let tmp_dir = TempDir::new()?;
-        let mut storage = SqliteStorage::new(tmp_dir.path(), true)?;
-        let uuid1 = Uuid::new_v4();
-        let uuid2 = Uuid::new_v4();
-        let uuid3 = Uuid::new_v4();
-
-        // create some operations
-        {
-            let mut txn = storage.txn()?;
-            txn.add_operation(Operation::Create { uuid: uuid1 })?;
-            txn.add_operation(Operation::Create { uuid: uuid2 })?;
-            txn.commit()?;
-        }
-
-        // read them back
-        {
-            let mut txn = storage.txn()?;
-            let ops = txn.operations()?;
-            assert_eq!(
-                ops,
-                vec![
-                    Operation::Create { uuid: uuid1 },
-                    Operation::Create { uuid: uuid2 },
-                ]
-            );
-
-            assert_eq!(txn.num_operations()?, 2);
-        }
-
-        // Clear them.
-        {
-            let mut txn = storage.txn()?;
-            txn.sync_complete()?;
-            txn.commit()?;
-        }
-
-        // create some more operations (to test adding operations after clearing)
-        {
-            let mut txn = storage.txn()?;
-            txn.add_operation(Operation::Create { uuid: uuid3 })?;
-            txn.add_operation(Operation::Delete {
-                uuid: uuid3,
-                old_task: TaskMap::new(),
-            })?;
-            txn.commit()?;
-        }
-
-        // read them back
-        {
-            let mut txn = storage.txn()?;
-            let ops = txn.operations()?;
-            assert_eq!(
-                ops,
-                vec![
-                    Operation::Create { uuid: uuid3 },
-                    Operation::Delete {
-                        uuid: uuid3,
-                        old_task: TaskMap::new()
-                    },
-                ]
-            );
-            assert_eq!(txn.num_operations()?, 2);
-        }
-
-        // Remove the wrong one
-        {
-            let mut txn = storage.txn()?;
-            assert!(txn
-                .remove_operation(Operation::Create { uuid: uuid3 })
-                .is_err())
-        }
-
-        // Remove the right one
-        {
-            let mut txn = storage.txn()?;
-            txn.remove_operation(Operation::Delete {
-                uuid: uuid3,
-                old_task: TaskMap::new(),
-            })?;
-            txn.commit()?;
-        }
-
-        // read the remaining op back
-        {
-            let mut txn = storage.txn()?;
-            let ops = txn.operations()?;
-            assert_eq!(ops, vec![Operation::Create { uuid: uuid3 },]);
-            assert_eq!(txn.num_operations()?, 1);
-        }
-
-        // Remove the last one
-        {
-            let mut txn = storage.txn()?;
-            txn.remove_operation(Operation::Create { uuid: uuid3 })?;
-            assert_eq!(txn.num_operations()?, 0);
-            txn.commit()?;
-        }
-
-        // Remove a nonexistent operation
-        {
-            let mut txn = storage.txn()?;
-            assert!(txn
-                .remove_operation(Operation::Create { uuid: uuid3 })
-                .is_err());
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_task_operations() -> Result<()> {
-        let tmp_dir = TempDir::new()?;
-        let mut storage = SqliteStorage::new(tmp_dir.path(), true)?;
-        let uuid1 = Uuid::new_v4();
-        let uuid2 = Uuid::new_v4();
-        let uuid3 = Uuid::new_v4();
-        let now = Utc::now();
-
-        // create some operations
-        {
-            let mut txn = storage.txn()?;
-            txn.add_operation(Operation::UndoPoint)?;
-            txn.add_operation(Operation::Create { uuid: uuid1 })?;
-            txn.add_operation(Operation::Create { uuid: uuid1 })?;
-            txn.add_operation(Operation::UndoPoint)?;
-            txn.add_operation(Operation::Delete {
-                uuid: uuid2,
-                old_task: TaskMap::new(),
-            })?;
-            txn.add_operation(Operation::Update {
-                uuid: uuid3,
-                property: "p".into(),
-                old_value: None,
-                value: Some("P".into()),
-                timestamp: now,
-            })?;
-            txn.commit()?;
-        }
-
-        // read them back
-        {
-            let mut txn = storage.txn()?;
-            let ops = txn.get_task_operations(uuid1)?;
-            assert_eq!(
-                ops,
-                vec![
-                    Operation::Create { uuid: uuid1 },
-                    Operation::Create { uuid: uuid1 },
-                ]
-            );
-            let ops = txn.get_task_operations(uuid2)?;
-            assert_eq!(
-                ops,
-                vec![Operation::Delete {
-                    uuid: uuid2,
-                    old_task: TaskMap::new()
-                }]
-            );
-            let ops = txn.get_task_operations(uuid3)?;
-            assert_eq!(
-                ops,
-                vec![Operation::Update {
-                    uuid: uuid3,
-                    property: "p".into(),
-                    old_value: None,
-                    value: Some("P".into()),
-                    timestamp: now,
-                }]
-            );
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn get_working_set_empty() -> Result<()> {
-        let tmp_dir = TempDir::new()?;
-        let mut storage = SqliteStorage::new(tmp_dir.path(), true)?;
-
-        {
-            let mut txn = storage.txn()?;
-            let ws = txn.get_working_set()?;
-            assert_eq!(ws, vec![None]);
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn add_to_working_set() -> Result<()> {
-        let tmp_dir = TempDir::new()?;
-        let mut storage = SqliteStorage::new(tmp_dir.path(), true)?;
-        let uuid1 = Uuid::new_v4();
-        let uuid2 = Uuid::new_v4();
-
-        {
-            let mut txn = storage.txn()?;
-            txn.add_to_working_set(uuid1)?;
-            txn.add_to_working_set(uuid2)?;
-            txn.commit()?;
-        }
-
-        {
-            let mut txn = storage.txn()?;
-            let ws = txn.get_working_set()?;
-            assert_eq!(ws, vec![None, Some(uuid1), Some(uuid2)]);
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn clear_working_set() -> Result<()> {
-        let tmp_dir = TempDir::new()?;
-        let mut storage = SqliteStorage::new(tmp_dir.path(), true)?;
-        let uuid1 = Uuid::new_v4();
-        let uuid2 = Uuid::new_v4();
-
-        {
-            let mut txn = storage.txn()?;
-            txn.add_to_working_set(uuid1)?;
-            txn.add_to_working_set(uuid2)?;
-            txn.commit()?;
-        }
-
-        {
-            let mut txn = storage.txn()?;
-            txn.clear_working_set()?;
-            txn.add_to_working_set(uuid2)?;
-            txn.add_to_working_set(uuid1)?;
-            txn.commit()?;
-        }
-
-        {
-            let mut txn = storage.txn()?;
-            let ws = txn.get_working_set()?;
-            assert_eq!(ws, vec![None, Some(uuid2), Some(uuid1)]);
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn set_working_set_item() -> Result<()> {
-        let tmp_dir = TempDir::new()?;
-        let mut storage = SqliteStorage::new(tmp_dir.path(), true)?;
-        let uuid1 = Uuid::new_v4();
-        let uuid2 = Uuid::new_v4();
-
-        {
-            let mut txn = storage.txn()?;
-            txn.add_to_working_set(uuid1)?;
-            txn.add_to_working_set(uuid2)?;
-            txn.commit()?;
-        }
-
-        {
-            let mut txn = storage.txn()?;
-            let ws = txn.get_working_set()?;
-            assert_eq!(ws, vec![None, Some(uuid1), Some(uuid2)]);
-        }
-
-        // Clear one item
-        {
-            let mut txn = storage.txn()?;
-            txn.set_working_set_item(1, None)?;
-            txn.commit()?;
-        }
-
-        {
-            let mut txn = storage.txn()?;
-            let ws = txn.get_working_set()?;
-            assert_eq!(ws, vec![None, None, Some(uuid2)]);
-        }
-
-        // Override item
-        {
-            let mut txn = storage.txn()?;
-            txn.set_working_set_item(2, Some(uuid1))?;
-            txn.commit()?;
-        }
-
-        {
-            let mut txn = storage.txn()?;
-            let ws = txn.get_working_set()?;
-            assert_eq!(ws, vec![None, None, Some(uuid1)]);
         }
 
         Ok(())
